feat: new cli for gt-node
