+++ packages/cli/src/cli/flags.ts
@@ -111,8 +111,13 @@ export function attachTranslateFlags(command: Command) {
   return command;
 }
 
-export function attachInlineTranslateFlags(command: Command) {
-  command
+/**
+ * Attaches flags necessary for parsing inline content
+ * @param command - The command to attach the flags to
+ * @returns The command with the inline content parsing flags attached
+ */
+function attachInlineContentParsingFlags(command: Command) {
+  return command
     .option(
       '--tsconfig, --jsconfig <path>',
       'Path to custom jsconfig or tsconfig file',
@@ -127,11 +132,35 @@ export function attachInlineTranslateFlags(command: Command) {
       '--inline',
       'Include inline content in translations (e.g., inline jsx translations, inline string translations, etc.)',
       true
+    );
+}
+
+/**
+ * Attaches flags necessary for validating a project
+ * @param command
+ * @returns The command with the validate flags attached
+ */
+export function attachValidateFlags(command: Command) {
+  return attachInlineContentParsingFlags(
+    command.option(
+      '-c, --config <path>',
+      'Filepath to config file, by default gt.config.json',
+      findFilepath(['gt.config.json'])
     )
-    .option(
+  );
+}
+
+/**
+ * Attaches flags necessary for translating a project
+ * @param command
+ * @returns The command with the translate flags attached
+ */
+export function attachInlineTranslateFlags(command: Command) {
+  return attachInlineContentParsingFlags(
+    command.option(
       '--ignore-errors',
       'Ignore errors encountered while scanning for inline content',
       false
-    );
-  return command;
+    )
+  );
 }
