+++ packages/cli/src/index.ts
@@ -3,6 +3,7 @@ import { NextCLI } from './cli/next.js';
 import { ReactCLI } from './cli/react.js';
 import { determineLibrary } from './fs/determineFramework.js';
 import { Command } from 'commander';
+import { NodeCLI } from './cli/node.js';
 
 export function main(program: Command) {
   program.name('gtx-cli');
@@ -13,6 +14,8 @@ export function main(program: Command) {
     cli = new NextCLI(program, library, additionalModules);
   } else if (library === 'gt-react') {
     cli = new ReactCLI(program, library, additionalModules);
+  } else if (library === 'gt-node') {
+    cli = new NodeCLI(program, library, additionalModules);
   } else {
     cli = new BaseCLI(program, library, additionalModules);
   }
